Tricky exception when reopening a campaign
If a campaign with one communication to one volunteer used prefix "A"
has been closed, next time volunteer will be triggered will also use
prefix "A". Thus, it should not be possible to reopen a campaign if
a prefix, from any volunteer and from any message of the communication
has been reused in a following campaign that is still open.